Modify central server to ensure pods are deleted correctly
Otherwise, take eg when pacman-chunk- 0 is being turned down. What would happen is that that pod would be deleted, and the replica count would decrease. However, then the highest number pod - eg pacman-chunk-3 - would be deleted and pacman-chunk-0 would be recreated. This ensures that does not happen and state is handled appropriately.