Integrate chunk 4 (#21)
* Modify central server to handle chunk 4

* Cleanup deploy action

* Update caddy-config to hopefully fix routing

* Switch to advanced stateful set to allow for appropriate routing

* Update logic using advanced stateful sets

* Update chunks to handle less specific ids

* formatting fix

* string over number?

* string it?

* Fix deployment for azure chunks

* Add logging

* Modify central server to ensure pods are deleted correctly

Otherwise, take eg when pacman-chunk- 0 is being turned down. What would happen is that that pod would be deleted, and the replica count would decrease. However, then the highest number pod - eg pacman-chunk-3 - would be deleted and pacman-chunk-0 would be recreated. This ensures that does not happen and state is handled appropriately.